Correct chapter (8-2. Is Z a shuffle of X and Y)
Change to correct chapter.  Z is said to be a shuffle of X and Y 
iff Z can be performed by interleaving the characters from X and Y in a way that maintains the left-to-right ordering of the characters 
from each string.   Give an efficient dynamic-programming algorithm that determines whether Z is a shuffle of X and Y.