Greedy & DP application of Knapsack variant (Coins)
Consider a country whose coins are minted with denominations of {d1,...,dk} units. We seek an algorithm to make change of n units using the minimum number of coins for this country. Show the greedy approach is not optimal then give an algo that is correct and analyze its run time.